% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_auc}
\alias{calculate_auc}
\title{Area under the curve (AUC) Calculation}
\usage{
calculate_auc(TP1, TP2, Time)
}
\arguments{
\item{TP1}{Value at the first timepoint}

\item{TP2}{Value at the second timepoint}

\item{Time}{Time interval in minutes}
}
\description{
Calculates the AUC between two timepoints using the trapezoid method
}
\examples{
data <- data \%>\% mutate(auc = calculate_auc(glucose_000, glucose_120, 120)
}
