% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_egfr}
\alias{calculate_egfr}
\title{Calculate eGFR}
\usage{
calculate_egfr(creatinine, age, sex, female = "female")
}
\arguments{
\item{creatinine}{Serum creatinine in mg/dl}

\item{age}{Age in year}

\item{sex}{Sex}

\item{female}{string of which value in sex equals female. Defaults to "female"}
}
\description{
Calculates the AUC between two timepoints using the trapezoid method
}
\examples{
data <- data \%>\% mutate(egfr = calculate_egfr(creatinine, age, sex))
}
