% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_fib4}
\alias{calculate_fib4}
\title{FIB-4 Calculation}
\usage{
calculate_fib4(age, ast, alt, platelets, unit = "conventional")
}
\arguments{
\item{age}{Age of the individual in years.}

\item{ast}{Aspartate aminotransferase (AST / GOT) level.}

\item{alt}{Alanine aminotransferase (ALT / GPT) level.}

\item{platelets}{Platelet count in 10^9/L.}

\item{unit}{Unit for AST/GOT and ALT/GPT. Use "conventional" for IU/L or "SI" for µkat/L. Defaults to "conventional".}
}
\value{
FIB-4 index
}
\description{
Calculates the FIB-4 index, a non-invasive score for assessing liver fibrosis.
}
\examples{
fib4 <- calculate_fib4(age = 55, ast = 45, alt = 30, platelets = 220, unit = "conventional")
data <- data \%>\% mutate(FIB4 = calculate_fib4(age, ast, alt, platelets))
}
