% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_FLI}
\alias{calculate_FLI}
\title{FLI Calculation}
\usage{
calculate_FLI(
  triglycerides,
  bmi,
  ggt,
  waist,
  unit_triglycerides = "conventional",
  unit_ggt = "conventional"
)
}
\arguments{
\item{triglycerides}{Triglycerides level.}

\item{bmi}{Body Mass Index (BMI).}

\item{ggt}{Gamma-glutamyl transferase (GGT) level.}

\item{waist}{Waist circumference in cm.}

\item{unit_triglycerides}{Triglycerides unit. Use "conventional" for mg/dl or "SI" for mmol/l. Defaults to "conventional".}

\item{unit_ggt}{GT unit. Use "conventional" for IU/L or "SI" for µkat/L. Defaults to "conventional".}
}
\value{
Fatty Liver Index (FLI)
}
\description{
Calculates the fatty liver index (FLI)
}
\examples{
FLI <- calculate_FLI(100, 24.2, 16, 70, unit_triglycerides = "conventional", unit_ggt = "conventional")
data <- data \%>\% mutate(FLI = calculate_FLI(triglycerides, bmi, ggt, waist, unit_triglycerides = "SI", unit_ggt = "conventional"))
}
