% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.r
\name{calculate_glycat}
\alias{calculate_glycat}
\title{Determine glycemic categories}
\usage{
calculate_glycat(glucose_000, glucose_120, hba1c, units = "mgdl")
}
\arguments{
\item{glucose_000}{The fasting glucose value}

\item{glucose_120}{The 120min glucose value}

\item{hba1c}{HbA1c value in %}

\item{units}{Unit of the glucose, can me mgdl or mmoll}
}
\value{
Glycemic Category
}
\description{
As Defined by: 
NGT -> Fasting glucose < 100, 120min glucose < 140
IFG: Fasting glucose elevated
IGT: 120min glucose elevated
IFG+IGT: Fasting ans 120min glucose elevated
DIA: Fasting glucose above 126 OR 120min glucose above 200 OR HBA1C above 6.5
}
\examples{
result <- calculate_glycat(90, 150, 5.8, units = "mgdl");
data <- data \%>\% mutate(category = calculate_glycat(glucose_000, glucose_120, hab1c))
}
