% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_HOMAB}
\alias{calculate_HOMAB}
\title{HOMA-B Calculation}
\usage{
calculate_HOMAB(insulin_000, glucose_000)
}
\arguments{
\item{insulin_000}{The fasting insulin value in pmol/l}

\item{glucose_000}{The fasting glucose value in mmol/l}
}
\value{
HOMAR-IR
}
\description{
Calculates the HOMA-B
}
\examples{
HOMAB <- HOMAB(40, 80);
data <- data \%>\% mutate(HOMAB = calculate_HOMAB(insulin_000, glucose_000))
}
