% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_HOMAIR}
\alias{calculate_HOMAIR}
\title{HOMA IR Calculation}
\usage{
calculate_HOMAIR(insulin_000, glucose_000)
}
\arguments{
\item{insulin_000}{The fasting insulin value in pmol/l}

\item{glucose_000}{The fasting glucose value in mmol/l}
}
\value{
HOMAR-IR
}
\description{
Calculates the HOMA-IR
}
\examples{
HOMAIR <- HOMAIR(40, 80);
data <- data \%>\% mutate(HOMAIR = calculate_HOMAIR(insulin_000,glucose_000))
}
