% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_OGIS}
\alias{calculate_OGIS}
\title{OGIS Calculation}
\usage{
calculate_OGIS(
  glucose_000,
  glucose_090,
  glucose_120,
  insulin_000,
  insulin_090,
  weight,
  height,
  dose = 75,
  unit_insulin = "SI",
  unit_glucose = "conventional"
)
}
\arguments{
\item{glucose_000}{The fasting glucose value.}

\item{glucose_090}{The 90 min glucose value.}

\item{glucose_120}{The 120 min glucose value.}

\item{insulin_000}{The fasting insulin value.}

\item{insulin_090}{The 90 min insulin value.}

\item{weight}{Weight in kg.}

\item{height}{Height in cm.}

\item{dose}{Amount of glucose in OGTT in g. Defaults to 75.}

\item{unit_insulin}{Character string specifying the unit for insulin. Use "SI" for pmol/l or "conventional" for µU/ml. Defaults to "SI".}

\item{unit_glucose}{Character string specifying the unit for glucose. Use "SI" for mmol/l or "conventional" for mg/dl. Defaults to "conventional".}
}
\value{
OGIS
}
\description{
Calculates the OGIS Index for a 5-point OGTT (Oral Glucose Insulin Sensitivity).
}
\examples{
# Example usage
data <- data \%>\% mutate(OGIS = calculate_OGIS(glucose_000, glucose_090, glucose_120, insulin_000, insulin_090, weight, height, unit_insulin = "SI", unit_glucose = "conventional"))
}
