% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{prep_data_km}
\alias{prep_data_km}
\title{prep_data_km}
\usage{
prep_data_km(data, date = "date", id = "id", event = "event")
}
\arguments{
\item{data}{Dataframe in long format with one row per timepoint.}

\item{date}{Name of the variable holding the measurement date. Defaults to "date".}

\item{id}{Name of the variable holding the patient ID. Defaults to "id".}

\item{event}{Name of the variable holding the event. Needs to be binary (0-1 or TRUE-FALSE). Defaults to "event".}
}
\value{
Dataframe with one row per id.
}
\description{
Function to bring longitudinal data into the appropriate format for survival analysis.
}
\examples{
test_km <- prep_data_km(data, id = "patno", date = "datum")
}
