% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{pvalue}
\alias{pvalue}
\title{Pvalue Calculation}
\usage{
pvalue(x, ...)
}
\description{
To be used in Table 1 package
Excludes "overall" column so that the p-value doesn't get calculated on that. Performs t-test for 2 variables, and anova for 3+ variable, chi square for categorical variables.
}
