% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{test_normality}
\alias{test_normality}
\title{Test Normality of Data}
\usage{
test_normality(x)
}
\arguments{
\item{x}{A numeric vector to test for normality.}
}
\value{
A character string indicating the recommended statistical test:
  "normal" if the data appears normally distributed (Shapiro-Wilk p > 0.05),
  or "non-normal" if the data does not appear normally distributed.
}
\description{
This function tests whether a numeric vector follows a normal distribution
and recommends an appropriate statistical test based on the result.
}
\examples{
# Test normally distributed data
normal_data <- rnorm(100)
test_normality(normal_data)

# Test non-normally distributed data
skewed_data <- rexp(100)
test_normality(skewed_data)

}
