% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_FLI}
\alias{calculate_FLI}
\title{FLI Calculation}
\usage{
calculate_FLI(triglycerides, bmi, ggt, waist)
}
\arguments{
\item{triglycerides}{Triglycerides in mg/dl}

\item{bmi}{BMI}

\item{ggt}{GGT levels in IU/L}

\item{waist}{Waist in cm}
}
\description{
Calculates the fatty liver index (FLI)
}
