% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_matsuda}
\alias{calculate_matsuda}
\title{Matsuda Calculation}
\usage{
calculate_matsuda(
  glucose_000,
  glucose_030,
  glucose_060,
  glucose_090,
  glucose_120,
  insulin_000,
  insulin_030,
  insulin_060,
  insulin_090,
  insulin_120
)
}
\arguments{
\item{glucose_000}{The fasting glucose value in mmol/l}

\item{glucose_030}{The fasting glucose value in mmol/l}

\item{glucose_060}{The fasting glucose value in mmol/l}

\item{glucose_090}{The 90min glucose value in mmol/l}

\item{glucose_120}{The 120min glucose value in mmol/l}

\item{insulin_000}{The fasting insulin value in pmol/l}

\item{insulin_030}{The fasting insulin value in pmol/l}

\item{insulin_060}{The fasting insulin value in pmol/l}

\item{insulin_090}{The fasting insulin value in pmol/l}

\item{insulin_120}{The 90min insulin value in pmol/l}
}
\value{
Matsuda
}
\description{
Calculates the Matsuda Index for a 5 point OGTT
}
\examples{
data <- data \%>\% mutate(Matsuda = calculate_matsuda(glucose_000, glucose_030, glucose_060, glucose_090, glucose_120, insulin_000, insulin_030, insulin_060, insulin_090, insulin_120))
}
