% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{calculate_OGIS}
\alias{calculate_OGIS}
\title{OGIS Calculation}
\usage{
calculate_OGIS(
  glucose_000,
  glucose_090,
  glucose_120,
  insulin_000,
  insulin_090,
  weight,
  height,
  dose = 75
)
}
\arguments{
\item{glucose_000}{The fasting glucose value in mmol/l}

\item{glucose_090}{The 90min glucose value in mmol/l}

\item{glucose_120}{The 120min glucose value in mmol/l}

\item{insulin_000}{The fasting insulin value in pmol/l}

\item{insulin_090}{The 90min insulin value in pmol/l}

\item{weight}{Weight in kg}

\item{height}{Height in cm}

\item{dose}{Amount of glucose in OGTT in g, defaults to 75}
}
\value{
OGIS
}
\description{
Calculates the OGIS Index for a 5 point OGTT
}
\examples{
data <- data \%>\% mutate(OGIS = calculate_OGIS(glucose_000, glucose_090, glucose_120, insulin_000, insulin_090, weight, height))
}
