% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifications.r
\name{classify_glycemia}
\alias{classify_glycemia}
\title{Determine glycemic categories}
\usage{
classify_glycemia(
  glucose_000,
  glucose_120,
  hba1c = NULL,
  system = "ADA",
  details = FALSE,
  unit_glucose = "conventional",
  unit_hba1c = "perc"
)
}
\arguments{
\item{glucose_000}{The fasting glucose value}

\item{glucose_120}{The 120min glucose value}

\item{hba1c}{HbA1c value. Not required when system = WHO.}

\item{system}{ADA or WHO. Defaults to ADA}

\item{details}{Logical (TRUE/FALSE). Wether the function should return a detailed differentiation of prediabetes or just "PRE". Defaults to FALSE}

\item{unit_glucose}{Unit of the glucose values, can be SI (mmol/l) or conventional (mg/dl). Defaults to conventional.}

\item{unit_hba1c}{Unit of the hba1c values, can be SI (mmol/mol) or perc (%). Defaults to perc.}
}
\value{
Glycemic Category
}
\description{
As Defined by:
NGT: Fasting glucose < 100, 120min glucose < 140, hba1c < 5.7 (ADA)
iIFG: Fasting glucose elevated
iIGT: 120min glucose elevated
IFG+IGT: Fasting and 120min glucose elevated
DIA: Fasting glucose above 126 OR 120min glucose above 200 OR HBA1C above 6.5
}
\examples{
classify_glycemia(90, 160, 6.9, system = "ADA", details = TRUE)
}
